fix(client): set nonce type to expected and existing type

+++ client/src/hooks/events/useAllocate.ts
@@ -27,7 +27,7 @@ const types = {
   ],
   AllocationPayload: [
     { name: 'allocations', type: 'Allocation[]' },
-    { name: 'nonce', type: 'integer' },
+    { name: 'nonce', type: 'uint256' },
   ],
 };
